backend/embedding_api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-genai>=1.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
redis>=5.0.0
SQLAlchemy>=2.0.0
